requirements.txt
requests
beautifulsoup4
sentence-transformers
faiss-cpu
fastapi
uvicorn
pydantic
openai
python-dotenv